# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: CtxOS API
  description: Context-driven OS Intelligence and Analysis API
  version: 1.0.0
  contact:
    name: CtxOS Team
    email: api@ctxos.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.ctxos.com
    description: Production server
...
paths:
...
  /api/v1/score:
    post:
      tags: [scoring]
      summary: Score entity
      description: Score an entity using specified scoring engines
      operationId: scoreEntity
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScoreRequest'
      responses:
        '200':
          description: Entity scored successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ScoringResult'
        '401':
          description: Unauthorized
        '403':
          description: Insufficient permissions
        '422':
          description: Validation error
...
  /api/v1/score/status:
    get:
      tags: [scoring]
      summary: Scoring service health
      description: Get scoring service health status
      operationId: scoringHealth
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Health status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthStatus'
  /api/v1/agents/run:
    post:
      tags: [agents]
      summary: Run agent
      description: Run a single agent on context data
      operationId: runAgent
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentRunRequest'
      responses:
        '200':
          description: Agent execution completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentResult'
        '404':
          description: Agent not found
  /api/v1/agents/pipeline:
    post:
      tags: [agents]
      summary: Run pipeline
      description: Run an agent pipeline
      operationId: runPipeline
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PipelineRunRequest'
      responses:
        '200':
          description: Pipeline execution completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PipelineResult'
  /api/v1/agents/bulk:
    post:
      tags: [agents]
      summary: Bulk analysis
      description: Run agents on multiple entities
      operationId: bulkAnalysis
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkAnalysisRequest'
      responses:
        '200':
          description: Bulk analysis completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkAnalysisResult'
  /api/v1/agents/list:
    get:
      tags: [agents]
      summary: List agents
      description: Get list of available agents
      operationId: listAgents
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Agents list retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentsList'
  /api/v1/agents/status/{agent_name}:
    get:
      tags: [agents]
      summary: Get agent status
      description: Get status of a specific agent
      operationId: getAgentStatus
      security:
        - bearerAuth: []
      parameters:
        - name: agent_name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Agent status retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentStatus'
        '404':
          description: Agent not found
  /api/v1/agents/pipelines:
    get:
      tags: [agents]
      summary: List pipelines
      description: Get list of available pipelines
      operationId: listPipelines
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Pipelines list retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PipelinesList'
  /api/v1/agents/create-pipeline:
    post:
      tags: [agents]
      summary: Create pipeline
      description: Create a new agent pipeline
      operationId: createPipeline
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePipelineRequest'
      responses:
        '200':
          description: Pipeline created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatePipelineResponse'
        '409':
          description: Pipeline already exists
...
  /api/v1/agents/audit-logs:
    get:
      tags: [agents]
      summary: Get audit logs
      description: Get agent operation audit logs
      operationId: getAgentAuditLogs
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Audit logs retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    HealthStatus:
      type: object
      required: [service, status, timestamp]
      properties:
        service:
          type: string
          description: Service name
        status:
          type: string
          enum: [healthy, degraded, unhealthy]
          description: Service health status
        timestamp:
          type: string
          format: date-time
          description: Health check timestamp
        features:
          type: object
          description: Service features
        statistics:
          type: object
          description: Service statistics
    PaginatedResponse:
      type: object
      required: [items, total, limit, offset, has_next]
      properties:
        items:
          type: array
          items:
            type: object
          description: Response items
        total:
          type: integer
          description: Total number of items
        limit:
          type: integer
          description: Items per page
        offset:
          type: integer
          description: Number of items skipped
        has_next:
          type: boolean
          description: Whether there are more items

    # Data Model Schemas
    EntityType:
      type: string
      enum: [host, domain, ip, user, service, application, network, database]
...
    SeverityLevel:
      type: string
      enum: [critical, high, medium, low, info]
    Context:
      type: object
      required: [entity, signals]
      properties:
        entity:
          $ref: '#/components/schemas/Entity'
        signals:
          type: array
          items:
            $ref: '#/components/schemas/Signal'
    Entity:
      type: object
      required: [id, entity_type, name]
      properties:
        id:
          type: string
          description: Entity ID
        entity_type:
          $ref: '#/components/schemas/EntityType'
          description: Entity type
        name:
          type: string
          description: Entity name
        description:
          type: string
          description: Entity description
        properties:
          type: object
          description: Entity properties
...
    ScoreRequest:
      type: object
      required: [context, engines]
      properties:
        context:
          $ref: '#/components/schemas/Context'
          description: Entity context
        engines:
          type: array
          items:
            type: string
          description: Scoring engines to use
        include_recommendations:
          type: boolean
          default: true
          description: Include recommendations
...
    ScoringResult:
      type: object
      required: [entity_id, entity_type, engine_name, score, severity, timestamp]
      properties:
        entity_id:
          type: string
          description: Entity ID
        entity_type:
          $ref: '#/components/schemas/EntityType'
          description: Entity type
        engine_name:
          type: string
          description: Scoring engine name
        score:
          type: number
          format: float
          description: Risk score (0-100)
        severity:
          $ref: '#/components/schemas/SeverityLevel'
          description: Risk severity
        factors:
          type: object
          description: Score factors
        recommendations:
          type: array
          items:
            type: string
          description: Recommendations
        timestamp:
          type: string
          format: date-time
          description: Scoring timestamp
...
    AgentRunRequest:
      type: object
      required: [agent_name, context]
      properties:
        agent_name:
          type: string
          description: Agent name
        context:
          $ref: '#/components/schemas/Context'
          description: Entity context
        scoring_result:
          type: object
          description: Scoring result context
        timeout_seconds:
          type: number
          format: float
          default: 30.0
          description: Execution timeout
    AgentResult:
      type: object
      required: [agent_name, success, duration_ms, timestamp]
      properties:
        agent_name:
          type: string
          description: Agent name
        success:
          type: boolean
          description: Execution success
        output:
          type: object
          description: Agent output
        error:
          type: string
          description: Error message
        duration_ms:
          type: number
          description: Execution duration in milliseconds
        timestamp:
          type: string
          format: date-time
          description: Execution timestamp
    PipelineRunRequest:
      type: object
      oneOf:
        - required: [pipeline_name, context]
        - required: [agents, context]
      properties:
        pipeline_name:
          type: string
          description: Pipeline name
        agents:
          type: array
          items:
            type: string
          description: Agent names for custom pipeline
        context:
          $ref: '#/components/schemas/Context'
          description: Entity context
        scoring_result:
          type: object
          description: Scoring result context
        parallel:
          type: boolean
          description: Run agents in parallel
        timeout_seconds:
          type: number
          format: float
          default: 60.0
          description: Pipeline timeout
    PipelineResult:
      type: object
      required: [pipeline_name, results, success_count, total_count, duration_ms, timestamp]
      properties:
        pipeline_name:
          type: string
          description: Pipeline name
        results:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/AgentResult'
          description: Agent results
        success_count:
          type: integer
          description: Number of successful agents
        total_count:
          type: integer
          description: Total number of agents
        duration_ms:
          type: number
          description: Pipeline duration in milliseconds
        timestamp:
          type: string
          format: date-time
          description: Execution timestamp

    # Configuration Schemas
...